@@ -1,10 +1,16 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Content Remove="wwwroot\fonts\IBMPlexSerif-Bold.ttf" />
+    <Content Remove="wwwroot\fonts\IBMPlexSerif-MediumItalic.ttf" />
+    <Content Remove="wwwroot\fonts\IBMPlexSerif-Regular.ttf" />
+  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="AutoMapper" Version="12.0.1" />
@@ -21,7 +27,20 @@
 
   <ItemGroup>
     <Folder Include="wwwroot\Coaches\" />
+    <Folder Include="wwwroot\fonts\" />
+    <Folder Include="wwwroot\Rooms\" />
     <Folder Include="wwwroot\Sport\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Update="Views\Admin\Delete.cshtml">
+      <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
+      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
+    </Content>
+    <Content Update="Views\Users\DeleteClientProfile.cshtml">
+      <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
+      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
+    </Content>
   </ItemGroup>
 
 </Project>
