@@ -9,15 +9,28 @@ namespace SportClub.BLL.DTO
     public class TrainingGroupDTO
     {
         public int Id { get; set; }
-        public int Number { get; set; }
-        public string Name { get; set; }
-        public int TimeId { get; set; }
-        public int RoomId { get; set; }
-        public string CoachName { get; set; }
-        public int? CoachId { get; set; }
-        public string GroupName { get; set; }
-        public int? GroupId { get; set; }
-        public string SpecialityName { get; set; }
-        public int? SpecialityId { get; set; }
+
+        public string Name { get; set; } = string.Empty;//название тренировки
+
+        //public int TimeId { get; set; }
+        public string? Time { get; set; }
+
+        public int Day { get; set; }
+        public string? DayName { get; set; }
+         
+
+        public string RoomName { get; set; } = string.Empty;
+        public int RoomId { get; set; }//id зала
+
+        public string CoachName { get; set; } = string.Empty;//имя тренера//
+        public int CoachId { get; set; }//?id тренера
+        public string? CoachPhoto { get; set; }//
+
+
+        //public string GroupName { get; set; } = string.Empty;//имя группы//
+        //public int GroupId { get; set; }//?id группы
+
+        public int Number { get; set; }//количество человек
+        public List<UserDTO> UsersId { get; set; } = new();//список человек
     }
 }
