@@ -11,19 +11,22 @@ namespace SportClub.BLL.DTO
         public int? Id { get; set; }
         public string? Name { get; set; }
        // public int TimeId { get; set; }
-       public string Time {  get; set; }
+        public string? Time {  get; set; }
         public int Day { get; set; }
-        public string DayName { get; set; }
+        public string? DayName { get; set; }
         public int RoomId { get; set; }
-        public string RoomName { get; set; }
+        public string? RoomName { get; set; }
         public string? CoachName { get; set; }
+        public string? CoachPhoto { get; set; }
         public int? CoachId { get; set; }
         public string? UserName { get; set; }
         public int? UserId { get; set; }
-       /// <summary>
-       //public string? SpecialityName { get; set; }
-       /// </summary>
-      //  public int? SpecialityId { get; set; }
+
+        //public string typeflag { get => "TrainingInd"; }
+        /// <summary>
+        //public string? SpecialityName { get; set; }
+        /// </summary>
+        //  public int? SpecialityId { get; set; }
         //  public int typeOfTrainingId { get; set; }
     }
 }
