@@ -20,6 +20,10 @@ namespace SportClub.BLL.DTO
         public string Login { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;      
         public List<int> trainingId { get; set; } = new();
-        public List<int> groupsId { get; set; } = new();
+
+        //public List<int> groupsId { get; set; } = new();
+
+        //public List<int> ScipShedulesId { get; } = new();
+        //public bool Recorded_by { get; set; } = false;
     }
 }
