@@ -16,7 +16,7 @@ namespace SportClub.DAL.Entities
        // public string Login { get; set; } = string.Empty;
       //  public string Password { get; set; } = string.Empty;
         public virtual List<TrainingInd>? trainingInds { get; set; }=new List<TrainingInd>();
-        public virtual List<Group> groups { get; set; }=new List<Group>() { };
+        
 
     }
 }
