@@ -9,10 +9,14 @@ namespace SportClub.DAL.Entities
     public class User : Person
     {
         public int Id { get; set; }
-       
-      //  public string Login { get; set; } = string.Empty;
-      //  public string Password { get; set; } = string.Empty;
-      //  public virtual List<TrainingInd> trainingInds { get; set; }
-        public virtual List<Group> groups { get; set; }=new List<Group>();
+
+        //  public string Login { get; set; } = string.Empty;
+        //  public string Password { get; set; } = string.Empty;
+        //  public virtual List<TrainingInd> trainingInds { get; set; }
+
+        public virtual List<TrainingGroup> traininggroups { get; set; } = new List<TrainingGroup>();
+        //public virtual List<Group> groups { get; set; } = new List<Group>();
+
+        //public bool Recorded_by { get; set; } = false;
     }
 }
