@@ -34,14 +34,14 @@ namespace SportClub.Controllers
         public IActionResult RegistrationClient()
         {
             HttpContext.Session.SetString("path", Request.Path);
-            return View("RegisterClient");
+            return View("RegisterClient"); 
         }
         public async Task<IActionResult> RegistrationCoach()
         {
             HttpContext.Session.SetString("path", Request.Path);
             await putSpecialities();
             await putPosts();
-            return View("RegisterCoach");
+            return View("RegisterCoach");/*RegisterCoach1*/
         }
         public async Task<IActionResult> RegistrationAdmin()
         {
@@ -69,6 +69,10 @@ namespace SportClub.Controllers
                     {
                         age--;
                     }
+                    if (age < 18)
+                    {
+                        ModelState.AddModelError("DateOfBirth", "Недопустимый возраст меньше 18лет");
+                    }
                 }
                 else
                 {
@@ -77,52 +81,23 @@ namespace SportClub.Controllers
             }
             catch { ModelState.AddModelError("DateOfBirth", "Некорректный формат даты рождения"); }
             if (ModelState.IsValid)
-            {
-                /* if (await userService.GetUser(user.Login) != null)
-                 {
-                     ModelState.AddModelError("login", "this login already exists");
-                     return View(user);
-                 }
-                 if (await userService.GetEmail(user.email) != null)
-                 {
-                     ModelState.AddModelError("email", "this email is already registred");
-                     return View(user);
-                 }*/
+            {             
                 AdminDTO u = new();
                 u.Login = user.Login;
                 u.Gender = user.Gender;
                 u.Email = user.Email;
                 u.Age = age;
                 u.Phone = user.Phone;
-                u.Name = user.Name;
-              //  u.Surname = user.Surname;
-              //  u.Dopname = user.Dopname;
+                u.Name = user.Name;            
                 u.DateOfBirth = user.DateOfBirth;
-
-                /*  byte[] saltbuf = new byte[16];
-                  RandomNumberGenerator randomNumberGenerator = RandomNumberGenerator.Create();
-                  randomNumberGenerator.GetBytes(saltbuf);
-                  StringBuilder sb = new StringBuilder(16);
-                  for (int i = 0; i < 16; i++)
-                      sb.Append(string.Format("{0:X2}", saltbuf[i]));
-                  string salt = sb.ToString();
-                  Salt s = new();
-                  s.salt = salt;
-                  string password = salt + user.Password;
-                  string hashedPassword = BCrypt.Net.BCrypt.HashPassword(password);
-                  u.Password = hashedPassword;*/
                 u.Password = user.Password;
+                u.Level = 1;
                 try
-                {
-                    /* db.Users.Add(u);
-                    db.SaveChanges();
-                    s.user = u;
-                     db.Salts.Add(s);
-                     db.SaveChanges();*/
+                {               
                    await adminService.AddAdmin(u);
                 }
                 catch { }
-                return RedirectToAction("Login");
+                return RedirectToAction("Index","Home");
             }
             return View("RegisterAdmin", user);
         }
@@ -136,7 +111,7 @@ namespace SportClub.Controllers
 
             try
             {
-                // DateTime dateTime = DateTime.Parse(user.DateOfBirth);
+                
                 DateTime birthDate;
                 if (DateTime.TryParse(user.DateOfBirth, out birthDate))
                 {
@@ -146,8 +121,10 @@ namespace SportClub.Controllers
                     {
                         age--;
                     }
-
-                    Console.WriteLine("Ваш возраст: " + age + " лет");
+                    if (age < 18)
+                    {
+                        ModelState.AddModelError("DateOfBirth", "Недопустимый возраст меньше 18лет");
+                    }
                 }
                 else
                 {
@@ -179,9 +156,10 @@ namespace SportClub.Controllers
                     u.Phone = user.Phone;
                     u.Photo = path;
                     u.Name = user.Name;
-                  //  u.Surname = user.Surname;
-                  //  u.Dopname = user.Dopname;
+                    //  u.Surname = user.Surname;
+                    //  u.Dopname = user.Dopname;
                     u.DateOfBirth = user.DateOfBirth;
+                    //u.DateOfBirth = user.DateOfBirth.ToString("dd.MM.yyyy");
                     u.Password = user.Password;
                     u.Description = user.Description;
                     u.PostId = user.PostId;
@@ -191,21 +169,21 @@ namespace SportClub.Controllers
                         await coachService.AddCoach(u);
                     }
                     catch { }
-                    return RedirectToAction("Login");
+                    return RedirectToAction("Index","Home");
                 }
             }
             await putSpecialities();
             await putPosts();
-            return View("RegisterCoach", user);
+            return View("RegisterCoach", user);/*RegisterCoach1*/
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> RegistrationClient(RegisterClientModel user)
         {
-            HttpContext.Session.SetString("path", Request.Path);
+           // HttpContext.Session.SetString("path", Request.Path);
 
             try
-            {
+            { 
                 DateTime birthDate;
                 if (DateTime.TryParse(user.DateOfBirth, out birthDate))
                 {
@@ -232,28 +210,31 @@ namespace SportClub.Controllers
                 u.Age = age;
                 u.Phone = user.Phone;
                 u.Name = user.Name;
-              //  u.Surname = user.Surname;
-              //  u.Dopname = user.Dopname;
-                u.DateOfBirth = user.DateOfBirth;
+                //  u.Surname = user.Surname;
+                //  u.Dopname = user.Dopname;
+                  u.DateOfBirth = user.DateOfBirth;
+                //u.DateOfBirth = user.DateOfBirth.ToString("dd.MM.yyyy");
                 u.Password = user.Password;
                 try
                 {
                     await userService.AddUser(u);
                 }
                 catch { return View("RegisterClient", user); }
-                return RedirectToAction("Login");
+
+                return RedirectToAction("Index", "Home");
+                //return RedirectToAction("Login");
             }
             return View("RegisterClient", user);
         }
         public IActionResult Login()
-        {        
+        {           
             return PartialView();
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Login(LoginModel user)
         {
-            HttpContext.Session.SetString("path", Request.Path);
+          //  HttpContext.Session.SetString("path", Request.Path);
 
             if (ModelState.IsValid)
             {
@@ -268,7 +249,7 @@ namespace SportClub.Controllers
                             HttpContext.Session.SetString("login", user.Login);
                             HttpContext.Session.SetString("admin", "admin");
                             HttpContext.Session.SetString("Id", a.Id.ToString());
-                            //return RedirectToAction("Index", "Home");
+                           // return RedirectToAction("Index", "Home");
                             return Json(true);
                         }
                         else
@@ -288,7 +269,7 @@ namespace SportClub.Controllers
                                 HttpContext.Session.SetString("login", user.Login);
                                 HttpContext.Session.SetString("coach", "coach");
                                 HttpContext.Session.SetString("Id", c.Id.ToString());
-                                //return RedirectToAction("Index", "Home");
+                               // return RedirectToAction("Index", "Home");
                                 return Json(true);
                             }
                             else
@@ -313,7 +294,7 @@ namespace SportClub.Controllers
                              HttpContext.Session.SetString("login", user.Login);                            
                              HttpContext.Session.SetString("client", "client");
                              HttpContext.Session.SetString("Id", u.Id.ToString());
-                        // return RedirectToAction("Index", "Home");
+                       //  return RedirectToAction("Index", "Home");
                         return Json(true);
                     }
                         else
@@ -341,6 +322,8 @@ namespace SportClub.Controllers
         [AcceptVerbs("Get", "Post")]
         public async Task<IActionResult> IsUserLoginInUse(string login)
         {
+            if (login == "admin")
+                return Json(false);
             UserDTO u = await userService.GetUserByLogin(login);         
             if (u == null)
                 return Json(true);
@@ -359,6 +342,8 @@ namespace SportClub.Controllers
         [AcceptVerbs("Get", "Post")]
         public async Task<IActionResult> IsCoachLoginInUse(string login)
         {
+            if (login=="admin")
+                return Json(false);
             CoachDTO c = await coachService.GetCoachByLogin(login);
             if (c == null)
                 return Json(true);
