@@ -11,6 +11,7 @@ namespace SportClub.Models
          ErrorMessageResourceName = "loginused")]
         public string? Login { get; set; }
 
+
         [Required(ErrorMessageResourceType = typeof(Resources.Resource),
                   ErrorMessageResourceName = "PassRequired")]
         [Display(Name = "password", ResourceType = typeof(Resources.Resource))]
@@ -18,12 +19,14 @@ namespace SportClub.Models
                  ErrorMessageResourceName = "LightPass")]
         public string? Password { get; set; }
 
+
         [Required(ErrorMessageResourceType = typeof(Resources.Resource),
                 ErrorMessageResourceName = "PassConRequired")]
         [Display(Name = "passwordConf", ResourceType = typeof(Resources.Resource))]
         [Compare("Password", ErrorMessageResourceType = typeof(Resources.Resource),
                 ErrorMessageResourceName = "passnoteq")]
         public string? PasswordConfirm { get; set; }
+
 
         [Required(ErrorMessageResourceType = typeof(Resources.Resource),
                 ErrorMessageResourceName = "EmailRequired")]
@@ -33,18 +36,26 @@ namespace SportClub.Models
                 ErrorMessageResourceName = "emailused")]
         public string? Email { get; set; }
 
+
         [Required(ErrorMessageResourceType = typeof(Resources.Resource), ErrorMessageResourceName = "Required")]
         [Display(Name = "DateB", ResourceType = typeof(Resources.Resource))]
+        [RegularExpression(@"^(0[1-9]|[12][0-9]|3[01]).(0[1-9]|1[012]).\d{4}$", ErrorMessage = "Введите дату в формате день/месяц/год")]
         public string? DateOfBirth { get; set; }
+         
+
 
         [Required(ErrorMessageResourceType = typeof(Resources.Resource), ErrorMessageResourceName = "Required")]
         [Display(Name = "Phone", ResourceType = typeof(Resources.Resource))]
-        [RegularExpression(@"\(\d{3}\)-\d{3} \d{2} \d{2}", ErrorMessage = "Введите номер в формате (097)-111 11 11")]
+        // [RegularExpression(@"\(\d{3}\)-\d{3} \d{2} \d{2}", ErrorMessage = "Введите номер в формате (097)-111 11 11")]
+        [RegularExpression(@"^(\+\d{1,3}\s?)?(\(\d{3}\)|\d{3}[-\.\s]?)(\d{3}[-\.\s]?)(\d{4})$", ErrorMessage = "Введите номер в формате +380971315143")]
+
         public string? Phone { get; set; }
+
 
         [Required(ErrorMessageResourceType = typeof(Resources.Resource), ErrorMessageResourceName = "Required")]
         [Display(Name = "gender", ResourceType = typeof(Resources.Resource))]
         public string Gender { get; set; }
+
 
         [Display(Name = "Name", ResourceType = typeof(Resources.Resource))]
         [Required(ErrorMessageResourceType = typeof(Resources.Resource), ErrorMessageResourceName = "Required")]
@@ -63,9 +74,11 @@ namespace SportClub.Models
         [Display(Name = "Post", ResourceType = typeof(Resources.Resource))]
         public int PostId { get; set; }
 
+
         [Required(ErrorMessageResourceType = typeof(Resources.Resource), ErrorMessageResourceName = "Required")]
         [Display(Name = "Speciality", ResourceType = typeof(Resources.Resource))]
         public int SpecialityId { get; set; }
+
 
         [Required(ErrorMessageResourceType = typeof(Resources.Resource), ErrorMessageResourceName = "Required")]
         [Display(Name = "Description", ResourceType = typeof(Resources.Resource))]
