@@ -68,7 +68,78 @@ namespace Resources {
                 return ResourceManager.GetString("Add", resourceCulture);
             }
         }
+        public static string startT
+        {
+            get
+            {
+                return ResourceManager.GetString("startT", resourceCulture);
+            }
+        }
+        public static string endT
+        {
+            get
+            {
+                return ResourceManager.GetString("endT", resourceCulture);
+            }
+        }
         
+             public static string saveSc
+        {
+            get
+            {
+                return ResourceManager.GetString("saveSc", resourceCulture);
+            }
+        }
+        public static string oDay
+        {
+            get
+            {
+                return ResourceManager.GetString("oDay", resourceCulture);
+            }
+        }
+        public static string addTtoCh
+        {
+            get
+            {
+                return ResourceManager.GetString("addTtoCh", resourceCulture);
+            }
+        }
+        public static string AllSc
+        {
+            get
+            {
+                return ResourceManager.GetString("AllSc", resourceCulture);
+            }
+        }
+        public static string chooseTime
+        {
+            get
+            {
+                return ResourceManager.GetString("chooseTime", resourceCulture);
+            }
+        }
+        public static string ChDay
+        {
+            get
+            {
+                return ResourceManager.GetString("ChDay", resourceCulture);
+            }
+        }
+
+        public static string SavT
+        {
+            get
+            {
+                return ResourceManager.GetString("SavT", resourceCulture);
+            }
+        }
+        public static string MakeCh
+        {
+            get
+            {
+                return ResourceManager.GetString("MakeCh", resourceCulture);
+            }
+        }
         /// <summary>
         ///   Looks up a localized string similar to добавить исполнителя.
         /// </summary>
@@ -569,7 +640,18 @@ namespace Resources {
                 return ResourceManager.GetString("Login", resourceCulture);
             }
         }
-        
+
+        /// <summary>
+        ///   Looks up a localized string similar to Войти.
+        /// </summary>
+        public static string LoginLength
+        {
+            get
+            {
+                return ResourceManager.GetString("LoginLength", resourceCulture);
+            }
+        }
+
         /// <summary>
         ///   Looks up a localized string similar to логин.
         /// </summary>
@@ -614,6 +696,7 @@ namespace Resources {
                 return ResourceManager.GetString("LogRequired", resourceCulture);
             }
         }
+     
         public static string man
         {
             get
@@ -624,7 +707,14 @@ namespace Resources {
         /// <summary>
         ///   Looks up a localized string similar to Музыкальный портал.
         /// </summary>
-        public static string MusicPortal {
+         public static string MakeTime
+        {
+            get
+            {
+                return ResourceManager.GetString("MakeTime", resourceCulture);
+            }
+         }
+public static string MusicPortal {
             get {
                 return ResourceManager.GetString("MusicPortal", resourceCulture);
             }
@@ -721,11 +811,19 @@ namespace Resources {
                 return ResourceManager.GetString("Post", resourceCulture);
             }
         }
+        public static string Profile
+        {
+            get
+            {
+                return ResourceManager.GetString("Profile", resourceCulture);
+            }
+        }
 
         /// <summary>
         ///   Looks up a localized string similar to Зарегестрироваться.
         /// </summary>
-        public static string Registration {
+        public static string Registration
+        {
             get {
                 return ResourceManager.GetString("Registration", resourceCulture);
             }
@@ -737,7 +835,20 @@ namespace Resources {
                 return ResourceManager.GetString("Required", resourceCulture);
             }
         }
-
+        public static string Save
+        {
+            get
+            {
+                return ResourceManager.GetString("Save", resourceCulture);
+            }
+        }
+        public static string Shedule
+        {
+            get
+            {
+                return ResourceManager.GetString("Shedule", resourceCulture);
+            }
+        }
         /// <summary>
         ///   Looks up a localized string similar to Название песни.
         /// </summary>
@@ -785,6 +896,13 @@ namespace Resources {
             get
             {
                 return ResourceManager.GetString("Surname", resourceCulture);
+            }
+        }
+        public static string Subcrip
+        {
+            get
+            {
+                return ResourceManager.GetString("Subcrip", resourceCulture);
             }
         }
 
@@ -875,6 +993,222 @@ namespace Resources {
         public static string YearSort {
             get {
                 return ResourceManager.GetString("YearSort", resourceCulture);
+            }
+        }
+
+
+        /// <summary>
+        ///   Looks up a localized string similar to Тренеры.
+        /// </summary>
+        public static string Coaches
+        {
+            get
+            {
+                return ResourceManager.GetString("Coaches", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Админы.
+        /// </summary>
+        public static string Admins
+        {
+            get
+            {
+                return ResourceManager.GetString("Admins", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Регистрация админа.
+        /// </summary>
+        public static string AdminRegistration
+        {
+            get
+            {
+                return ResourceManager.GetString("AdminRegistration", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Регистрация клиента.
+        /// </summary>
+        public static string UserRegistration
+        {
+            get
+            {
+                return ResourceManager.GetString("UserRegistration", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Регистрация тренера.
+        /// </summary>
+        public static string CoachRegistration
+        {
+            get
+            {
+                return ResourceManager.GetString("CoachRegistration", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Регистрация пункт меню.
+        /// </summary>
+        public static string RegistrationM
+        {
+            get
+            {
+                return ResourceManager.GetString("RegistrationM", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Должности/Специальности.
+        /// </summary>
+        public static string PositionsAndSpecialties
+        {
+            get
+            {
+                return ResourceManager.GetString("PositionsAndSpecialties", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Время/график.
+        /// </summary>
+        public static string TimeAndSchedule
+        {
+            get
+            {
+                return ResourceManager.GetString("TimeAndSchedule", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Создать время
+        /// </summary>
+        public static string CreateTime
+        {
+            get
+            {
+                return ResourceManager.GetString("CreateTime", resourceCulture);
+            }
+        }
+
+
+        /// <summary>
+        ///   Looks up a localized string similar to Создать график дня
+        /// </summary>
+        public static string Create_A_ScheduleForDay
+        {
+            get
+            {
+                return ResourceManager.GetString("Create_A_ScheduleForDay", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Добавление/Редактирование залов
+        /// </summary>
+        public static string AddingAndEditingHall
+        {
+            get
+            {
+                return ResourceManager.GetString("AddingAndEditingHall", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Зал
+        /// </summary>
+        public static string Rooms
+        {
+            get
+            {
+                return ResourceManager.GetString("Rooms", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Мои тренировки
+        /// </summary>
+        public static string MyWorkouts
+        {
+            get
+            {
+                return ResourceManager.GetString("MyWorkouts", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Время начала
+        /// </summary>
+        public static string TimeStart
+        {
+            get
+            {
+                return ResourceManager.GetString("TimeStart", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Время окончания
+        /// </summary>
+        public static string TimeEnd
+        {
+            get
+            {
+                return ResourceManager.GetString("TimeEnd", resourceCulture);
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Удалить время
+        /// </summary>
+        public static string DeleteTime
+        {
+            get
+            {
+                return ResourceManager.GetString("DeleteTime", resourceCulture);
+            }
+        }
+        
+
+        /// <summary>
+        ///   Looks up a localized string similar to Сохраненное время
+        /// </summary>
+        public static string SavedTime
+        {
+            get
+            {
+                return ResourceManager.GetString("SavedTime", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Создать время для графиков
+        /// </summary>
+        public static string AddTimeForCharts
+        {
+            get
+            {
+                return ResourceManager.GetString("AddTimeForCharts", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to График дня
+        /// </summary>
+        public static string GraphOfDay
+        {
+            get
+            {
+                return ResourceManager.GetString("GraphOfDay", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Выход
+        /// </summary>
+        public static string Exit
+        {
+            get
+            {
+                return ResourceManager.GetString("Exit", resourceCulture);
             }
         }
     }
