@@ -1,14 +1,16 @@
-﻿<form asp-controller="Login" method="post">
+﻿<form asp-controller="Admin" method="post">
     @model SportClub.BLL.DTO.PostDTO
-
-    <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-    <div class="form-group">
-        <span>@Resources.Resource.Name </span><br />
+   @*  <h2>Потвердите удаление должности тренера:</h2> *@
+   @*  <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+    <div class="div_reg padd"> *@
+    <span style="color:white">удаление должности тренера:</span><br />
         <input asp-for="Name" class="form-control card _c7" readonly />
         <input asp-for="Id" type="hidden" />
         <span asp-validation-for="Name" class="text-danger"></span>
-    </div>
-    <div class="form-group">
-        <input type="submit" formaction="EditSpeciality" value="@Resources.Resource.Delete" class="btn btn-primary c2" />
-    </div>
+  @* 
+        <br />
+        <a asp-area="" asp-controller="Admin" asp-action="BackToPost" class="btn btn-primary buttons1">Назад</a>
+        <br />
+        <input type="submit" formaction="ConfirmDeletePost" value="@Resources.Resource.Delete" class="btn btn-primary red" />
+    </div> *@
 </form>
