@@ -3,7 +3,135 @@
 @{
     ViewData["Title"] = "CoachProfile";
 }
-<img src=@Model.Photo width="100" height="100" alt="photo" />
+
+@using Microsoft.AspNetCore.Http
+<div style="background-color: #374566; box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black; border-radius: 15px; width: 500px; margin-left: 400px;">
+    <p class="font" style="font-size: 32px; color: white; padding-top: 30px; text-align: center">
+        <img src=@Model.Photo width="130" height="130" alt="photo1" style="padding:20px;margin-left:30px" id="preview" />Мой профиль
+    </p>
+    <hr style="color:white" />
+    <div class="row">
+        <div>
+            <br />
+            <form method="post" enctype="multipart/form-data">
+                <input type="hidden" asp-for="Id" />
+                <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+                @Html.AntiForgeryToken()
+                <div>
+                   
+                    <div class="form-group" style="margin-left: 40px;">
+                        <span class="white-color_size20 font">Изменить фото</span><br />               
+                        <input class="white-color_size20" type="file" name="p" accept=".jpg, .jpeg, .png" onchange="previewImage(event)" /><br /><br />
+                    </div>
+                </div>
+                <div class="row">
+                    @Html.ValidationSummary(true, "")
+                    <div style="width: 500px;  justify-content: center; ">
+
+                        <div class="form-group" style="margin-left: 40px"> 
+                            <label asp-for="Name" class="control-label" style="color: white" class="font"></label>
+                            <input asp-for="Name" type="text" class="form-control" name="name" id="name" class="font" style="width: 420px">
+                            <span asp-validation-for="Name" class="text-danger" style="width: 450px"></span>
+                        </div>
+                        <div class="form-group" style="margin-left: 40px">
+                             
+                            <label asp-for="Login" class="control-label" style="color: white" class="font"></label>
+                            <input asp-for="Login" type="text" class="form-control" name="login" id="login" style="width: 420px">
+                            <span asp-validation-for="Login" class="text-danger" style="width: 450px"></span>
+                        </div>
+                       
+                        <div class="form-group" style="margin-left: 40px">
+                            <label asp-for="Email" class="control-label" style="color: white" class="font"></label>
+                            <input asp-for="Email"  class="form-control"  id="emailInput" style="width: 420px">
+                            <span asp-validation-for="Email" class="text-danger" style="width: 450px"></span>
+                        </div>                    
+                        
+                        <div class="form-group" style="margin-left: 40px">
+                            <label asp-for="Phone" style="color: white" class="font">@Resources.Resource.Phone</label>
+                            <input asp-for="Phone" class="form-control" placeholder="(097)-111 11 11" style="width: 420px" />
+                            <span asp-validation-for="Phone" class="text-danger" style="width: 450px"></span>
+                        </div>
+                        <div class="form-group" style="margin-left: 40px">
+                            <label style="color: white" class="font">@Resources.Resource.gender</label>
+                            <div class="form-check">
+                                <input class="form-check-input" type="radio" asp-for="Gender" value="Мужской" id="male" />
+                                <label style="color: white" class="font form-check-label" for="male">@Resources.Resource.man</label>
+                            </div>
+                            <div class="form-check">
+                                <input class="form-check-input" type="radio" asp-for="Gender" value="Женский" id="female" />
+                                <label style="color: white" class="font form-check-label" for="female">@Resources.Resource.female</label>
+                            </div>
+                        </div>
+                        <div class="form-group" style="margin-left: 40px">
+                            <label style="color: white" class="font">@Resources.Resource.DateB</label><br /> 
+                            <input asp-for="DateOfBirth" type="text"  />
+                            <br />
+                            <span asp-validation-for="DateOfBirth" class="text-danger" style="width: 450px"></span>
+                        </div>                       
+                        <br />
+                        <br />
+                        
+                        <div style="margin-left: 30px;padding-bottom:20px;">
+                            <input type="submit" class="btn btn-outline-light" formaction="EditCoachProfile" value="Редактировать" />
+                            <input class="btn btn-outline-light" formaction="ChangeCoachPassword" type="submit" value="Изменить пароль" />
+                           @*  <input type="submit" class="btn btn-outline-light" formaction="DeleteCoachProfile" value="Удалить и Выйти" /> *@
+                        </div>
+                    </div>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+@section Scripts {
+    @{
+        await Html.RenderPartialAsync("_ValidationScriptsPartial");
+    }
+}
+<script>
+    $(document).ready(function () {
+
+        $("#login").on("change", function (e) {
+            let log = e.target.value;
+            let r = true;
+            $.getJSON('/Login/IsLoginIn', function (data) {
+                $.each(data, function (key, val) {
+                    if (log == val) {
+                        $("#log").css("display", "block");
+                        r = false;
+                    }
+                    if (r) {
+                        $("#log").css("display", "none");
+                    }
+                }) 
+            }); 
+        });
+
+
+        $('#emailInput').on('blur', function () {
+            var email = $(this).val();
+            $.get('/Login/IsEmailInUse', { email: email }, function (data) {
+                if (data === true) {
+                    $('#emailStatus').text('Email уже занят.');
+                } 
+            });
+        });
+    });
+
+    function previewImage(event) {
+        var input = event.target;
+        var preview = document.getElementById('preview');
+
+        var reader = new FileReader();
+
+        reader.onload = function () {
+            preview.src = reader.result;
+        };
+
+        reader.readAsDataURL(input.files[0]);
+    }
+</script>
+
+@* <img src=@Model.Photo width="100" height="100" alt="photo" />
 <form method="get" enctype="multipart/form-data">
     @Html.AntiForgeryToken()
     <div class="card c22">
@@ -20,7 +148,7 @@
             @Html.DisplayNameFor(model => model.Login)
             <div>
                 @Html.EditorFor(model => model.Login, new { id = "login" })
-                @*<span id="log" style="color:red; display:none">@Resources.Resource.loginused</span>*@
+                @*<span id="log" style="color:red; display:none">@Resources.Resource.loginused</span>@*
                 <span asp-validation-for="Login" class="text-danger"></span>
             </div>
         </div>
@@ -79,4 +207,4 @@
     @{
         await Html.RenderPartialAsync("_ValidationScriptsPartial");
     }
-}
+} *@
