@@ -4,8 +4,65 @@
 @{
     ViewData["Title"] = "Details";
 }
+<div class="div_reg">
+    <p class="font" style="font-size: 32px; color: white; padding-top: 30px; text-align: center">
+        Details coach
+    </p>
 
-<h1>Details</h1>
+    <div class="row">
+        <div>
+
+            <div class="row">
+                <div >
+                    <img src=@Model.Photo width="130" height="130" alt="photo1" style="padding:20px;margin-left:30px" />
+                </div>
+                <div style="width: 500px;  justify-content: center; ">
+                    <div class="form-group" style="margin-left: 40px">
+                        <label asp-for="Name" class="control-label font  white-color_size20" style="color: white"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Name)</h1>
+                    </div>
+                    <div class="form-group" style="margin-left: 40px">
+
+                        <label asp-for="Login" class="control-label font  white-color_size20" style="color: white"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Login)</h1>
+                    </div>
+
+                    <div class="form-group" style="margin-left: 40px">
+                        <label asp-for="Email" class="control-label font  white-color_size20" style="color: white"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Email)</h1>
+                    </div>
+
+                    <div class="form-group" style="margin-left: 40px">
+                        <label asp-for="Phone" style="color: white" class="control-label font  white-color_size20"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Phone)</h1>
+                    </div>
+                </div>
+
+                <div class="form-group" style="margin-left: 40px">
+                    <label asp-for="Gender" class="control-label font  white-color_size20" style="color: white"></label>
+                    <h1 class="form-control p1" style="width: 420px;height: 33px">@Html.DisplayFor(model => model.Gender)</h1>
+                </div>
+
+                <div class="form-group" style="margin-left: 40px">
+                    <label asp-for="DateOfBirth" class="control-label font  white-color_size20" style="color: white"></label>
+                    <h1 class="form-control p1" style="width: 420px;height: 33px">@Html.DisplayFor(model => model.DateOfBirth)</h1>
+                </div> <br />
+                <div class="form-group" style="margin-left: 40px">
+                    <label class="font white-color_size20">@Resources.Resource.Description</label>
+                    <textarea asp-for="Description" style="width: 420px" readonly></textarea>
+                    <span asp-validation-for="Description" class="text-danger" style="width: 450px"></span><br />
+                </div><br />
+                <div style="margin: 10px auto 25px auto;width: max-content;">
+                    <a class="btn btn-outline-light" asp-controller="Coach" asp-action="Edit" asp-route-id="@Model?.Id">Edit</a>
+                    <a class="btn btn-outline-light" asp-action="GetCoaches" asp-controller="Coach">Back to List</a>
+                </div>
+                <br /><br />
+            </div>
+
+        </div>
+    </div>
+</div>
+@* <h1>Details</h1>
 
 <div>
     <h4>Coach</h4>
@@ -72,3 +129,4 @@
   
 </div>
 
+ *@
