@@ -1,126 +1,116 @@
-﻿@*
-    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-*@
-@model SportClub.Models.RegisterCoachModel
+﻿@model SportClub.Models.RegisterCoachModel
 @using Microsoft.AspNetCore.Http
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <link href='https://fonts.googleapis.com/css?family=Roboto:400,100,300,700' rel='stylesheet' type='text/css'>
-
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-
-    <link rel="stylesheet" href="css/style.css">
-
-</head>
-<body>
-<div style="background-color: #374566; box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black; border-radius: 15px; width: 500px; margin-left: 400px;">
-     <p class="font" style="font-size: 32px; color: white; padding-top: 30px; text-align: center">Регистрация тренера</p>
-     <hr style="color:white"/>
-     <div class="row">
-     <div >
-        <br />
-    <form id="registrationCoachForm" action="RegistrationCoach" method="post" enctype="multipart/form-data" >
-         <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-           @Html.AntiForgeryToken()
-           <div class="row">
+ 
+<div class="div_reg">
+    <p class="font">@Resources.Resource.CoachRegistration</p>@* Регистрация тренера *@
+    <div class="row">
+        <form id="registrationCoachForm" action="RegistrationCoach" method="post" enctype="multipart/form-data">
+            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+            @Html.AntiForgeryToken()
             @Html.ValidationSummary(true, "")
-            <div style="width: 500px;  justify-content: center; ">
 
-					<div class="form-group" style="margin-left: 40px">
-					   @*  <div class="font form-control" style="width: 420px; color: white; border-color: none" >
-                        @Html.EditorFor(model => model.Name,
-                        new {AddHtmlAttributeValue = new {@class = "form-control"}})
-                        </div> *@
-                        <label asp-for="Name" class="control-label" style="color: white" class="font"></label>
-						<input asp-for="Name" type="text" class="form-control" name="name" id="name" class="font" style="width: 420px">
-                        <span asp-validation-for="Name" class="text-danger" style="width: 450px"></span>
-					</div>        
-					<div class="form-group" style="margin-left: 40px">
-                        @* <div class="font form-control" style="width: 420px; color: white; border-color: none" >
-                        @Html.EditorFor(model => model.Login)
-                        </div> *@
-                        <label asp-for="Login" class="control-label"style="color: white" class="font"></label>
-						<input asp-for="Login" type="text" class="form-control" name="login" id="login" style="width: 420px">
-                        <span asp-validation-for="Login" class="text-danger" style="width: 450px"></span>
-					</div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label asp-for="Password" class="control-label" style="color: white" class="font"></label>
-						<input asp-for="Password" type="password" class="form-control" name="password" id="password" style="width: 420px">
-                        <span asp-validation-for="Password" class="text-danger" style="width: 450px"></span>
-					</div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label asp-for="PasswordConfirm" class="control-label" style="color: white; width: 200px" class="font"></label>
-						<input asp-for="PasswordConfirm" type="password" class="form-control" name="passwordconf" id="password" style="width: 420px">
-                        <span asp-validation-for="PasswordConfirm" class="text-danger" style="width: 450px"></span>
-					</div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label asp-for="Email" class="control-label" style="color: white" class="font"></label>
-						<input asp-for="Email" type="email" class="form-control" name="password" id="emailInput" style="width: 420px">
-                        <span asp-validation-for="Email" class="text-danger" style="width: 450px"></span>
-					</div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <span style="color: white" class="font">@Resources.Resource.Photo</span><br />
-                        <input style="color: white" type="file" name="p" accept=".jpg, .jpeg" /><br /><br />
-                    </div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label style="color: white" class="font">@Resources.Resource.Post</label>
-                        <select asp-for="PostId" class="form-control" asp-items="ViewBag.PostId" style="width: 420px"></select>
-                        <span asp-validation-for="PostId" class="text-danger"></span>
-                    </div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label style="color: white" class="font">@Resources.Resource.Speciality</label>
-                        <select asp-for="SpecialityId" class="form-control" asp-items="ViewBag.SpecialityId" style="width: 420px"></select>
-                        <span asp-validation-for="SpecialityId" class="text-danger" style="width: 450px"></span>
-                    </div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label asp-for="Phone" style="color: white" class="font">@Resources.Resource.Phone</label>
-                        <input asp-for="Phone" class="form-control" placeholder="(097)-111 11 11" style="width: 420px"/>
-                        <span asp-validation-for="Phone" class="text-danger" style="width: 450px"></span>
-                    </div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label style="color: white" class="font">@Resources.Resource.gender</label>
-                        <div class="form-check">
-                            <input class="form-check-input" type="radio" asp-for="Gender" value="Мужской" id="male" />
-                            <label style="color: white" class="font" class="form-check-label" for="male">@Resources.Resource.man</label>
-                        </div>
-                        <div class="form-check">
-                            <input class="form-check-input" type="radio" asp-for="Gender" value="Женский" id="female" />
-                            <label style="color: white" class="font" class="form-check-label" for="female">@Resources.Resource.female</label>
-                        </div>
-                    </div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label style="color: white" class="font">@Resources.Resource.DateB</label><br />
-                        <input type="date" asp-for="DateOfBirth" /><br />
-                        <span asp-validation-for="DateOfBirth" class="text-danger" style="width: 450px"></span>
-                    </div>
-                    <div class="form-group" style="margin-left: 40px">
-                        <label style="color: white" class="font">@Resources.Resource.Description</label>
-                        <textarea asp-for="Description" style="width: 420px"></textarea>
-                        <span asp-validation-for="Description" class="text-danger" style="width: 450px"></span><br />
-                    </div>
-                    <br />
-                    <br />
-                    <div class="form-group" style="margin-left: 200px">
-                        <button type="submit" class="btn btn-primary font y" style="color: black">@Resources.Resource.Confirm</button><br /><br />
-                    </div>
+            <div class="form-group" style="margin-left: 40px">
+
+                <label asp-for="Name" class="control-label font white-color_size20"></label>
+                <input asp-for="Name" type="text" class="form-control font" style="width: 420px">
+                <span asp-validation-for="Name" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+
+                <label asp-for="Login" class="control-label font white-color_size20"></label>
+                <input asp-for="Login" type="text" class="form-control" id="login" style="width: 420px" autocomplete="off">
+                <span asp-validation-for="Login" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label asp-for="Password" class="control-label font white-color_size20"></label>
+                <input asp-for="Password" type="password" class="form-control" style="width: 420px" autocomplete="off">
+                <span asp-validation-for="Password" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label asp-for="PasswordConfirm" style=" width: 200px" class="control-label font white-color_size20"></label>
+                <input asp-for="PasswordConfirm" type="password" class="form-control" style="width: 420px">
+                <span asp-validation-for="PasswordConfirm" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label asp-for="Email" class="control-label font white-color_size20"></label>
+                <input asp-for="Email" type="email" class="form-control" id="emailInput" style="width: 420px">
+                <span asp-validation-for="Email" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px;">
+                <span class="white-color_size20 font">@Resources.Resource.Photo</span><br />
+                <input class="white-color_size20" type="file" name="p" accept=".jpg, .jpeg, .png" /><br /><br />
 
             </div>
-        </div>
-    </form>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label class="font white-color_size20">@Resources.Resource.Post</label>
+                <select asp-for="PostId" class="form-control" asp-items="ViewBag.PostId" style="width: 420px"></select>
+                <span asp-validation-for="PostId" class="text-danger"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label class="font white-color_size20">@Resources.Resource.Speciality</label>
+                <select asp-for="SpecialityId" class="form-control" asp-items="ViewBag.SpecialityId" style="width: 420px"></select>
+                <span asp-validation-for="SpecialityId" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label asp-for="Phone" class="font white-color_size20">@Resources.Resource.Phone</label>
+                <input asp-for="Phone" class="form-control" placeholder="+380973456789" style="width: 420px" />
+                <span asp-validation-for="Phone" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label class="font white-color_size20">@Resources.Resource.gender</label>
+                <div class="form-check">
+                    <input class="form-check-input" type="radio" asp-for="Gender" value="Мужской" id="male" />
+                    <label class="font form-check-label white-color_size20" for="male">@Resources.Resource.man</label>
+                </div>
+                <div class="form-check">
+                    <input class="form-check-input" type="radio" asp-for="Gender" value="Женский" id="female" />
+                    <label class="font form-check-label white-color_size20" for="female">@Resources.Resource.female</label>
+                </div>
+                <span asp-validation-for="Gender" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label class="font white-color_size20">@Resources.Resource.DateB</label><br />
+                <input class="form-control font" style="width: 420px" asp-for="DateOfBirth" />
+                <span asp-validation-for="DateOfBirth" class="text-danger" style="width: 450px"></span>
+            </div>
+
+            <div class="form-group" style="margin-left: 40px">
+                <label class="font white-color_size20">@Resources.Resource.Description</label>
+                <textarea asp-for="Description" style="width: 420px"></textarea>
+                <span asp-validation-for="Description" class="text-danger" style="width: 450px"></span><br />
+            </div>
+
+            @* <br />
+            <div class="form-group" style="margin-left: 200px">
+                <button type="submit" class="btn btn-primary font y">@Resources.Resource.Confirm</button><br /><br />
+            </div> *@
+
+            <div style="margin: 10px auto 10px auto;width: max-content;">
+                <button style="height: 38px;" type="submit" class="btn btn-primary font y">@Resources.Resource.Confirm</button>
+                <a asp-action="Index" asp-controller="Home" style="margin-left: 10px;" class="btn btn-outline-light">@Resources.Resource.Back</a><br />@* Back *@
+            </div><br />
+        </form>
+    </div>
 </div>
-</div>
-</div>
-@section Scripts {
-    @{
-        await Html.RenderPartialAsync("_ValidationScriptsPartial");
+
+    
+    @section Scripts {
+        @{
+            await Html.RenderPartialAsync("_ValidationScriptsPartial");
+        }
     }
-}
-<script>
-    $(document).ready(function () {
+    <script>
+        $(document).ready(function () {
 
         $("#login").on("change", function (e) {
             let log = e.target.value;
@@ -135,11 +125,7 @@
                         $("#log").css("display", "none");
                     }
                 })
-
-
             });
-
-
         });
 
 
@@ -149,10 +135,7 @@
                 if (data === true) {
                     $('#emailStatus').text('Email уже занят.');
                 }
-
             });
         });
     });
 </script>
-</body>
-</html>
