@@ -1,41 +1,404 @@
-﻿<h4 class=" b1 b3 c20 center pr">Время</h4>
-<div class="center pr">
-    <div class="flex">
+﻿@Html.AntiForgeryToken()
 
-        <div class=" card c7">
-            <h4>@Resources.Resource.Addition</h4>
-            <form asp-controller="Time" method="post">
-                @model SportClub.BLL.DTO.TimeTDTO
+<div class="flex" style="justify-content:space-around;">
+    <div >
+        <a asp-area="" asp-controller="Time" asp-action="Exit" class="btn btn-primary buttonsGY">X</a>
+    </div>
+   <div >
+        <h4 class="whiteY" style="text-align:center;padding:5px; box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black;">
+            @Resources.Resource.MakeTime
+        </h4>
+        @model SportClub.Models.TimeTimetableModel
 
+  
+     <div class="flex padd div_reg2">
+
+            <div class=" card c7 padd" style=" background-color: #2F4F4F;padding-top:0px;">
+                <h4 class="whiteY" style="text-align:center;">@Resources.Resource.Addition</h4>
+                <form asp-controller="Time" method="post" style="border-radius:3px; background-color: white;padding:10px;border:0px;">
+               @*  @model SportClub.BLL.DTO.TimeTDTO
+                @Html.AntiForgeryToken() *@
+             
                 <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-                <div class="form-group">
-                    <span>Время начала</span><br />
-                    <input name="Start" class="form-control card _c7" />
-                    <span asp-validation-for="StartTime" class="text-danger"></span>
+                    <div class="form-group" >
+                        <span style="text-align:center;">@Resources.Resource.startT</span><br />
+                        <input name="Start" id="start" class="form-control card _c7" placeholder="00:00" style=" background-color: white" />
+                    <span  id="st" class="text-danger"></span>
                 </div>
                 <div class="form-group">
-                    <span>Время окончания</span><br />
-                    <input name="End" class="form-control card _c7" />
-                    <span asp-validation-for="EndTime" class="text-danger"></span>
-                </div>
-                <div class="form-group">
-                    <input type="submit" formaction="AddTimeT" value="@Resources.Resource.Add" class="btn btn-primary c2" />
-                </div>
+                        <span style="text-align:center;">@Resources.Resource.endT</span><br />
+                        <input name="End" id="end" class="form-control card _c7" placeholder="00:00" style=" background-color: white" />
+                    <span  id="en" class="text-danger"></span>
+
+                 </div>
+                    <br /><br /><br />
+                    <div class="form-group"  style=" background-color: #2F4F4F;padding:20px;margin:-10px;justify-content:center;">
+                        <input type="submit" id="add" formaction="AddTimeT" value="@Resources.Resource.Add" class="btn btn-primary  buttonsGY" />
+                 </div>
             </form>
 
         </div>
-        <div class=" card c7">
-            <h4> @Resources.Resource.Editing</h4>
-            <form asp-controller="Time" method="post">
-                <label>График</label>
-                <select asp-for="Id" size="14" class="form-control card _c7" asp-items="ViewBag.TimeId"></select>
+        <div style="width:20px"></div>
+            <div class=" card c7 padd" style=" background-color: #2F4F4F;padding-top:0px;">
 
-                <br />
-                <div class="form-group">           
-                    <input type="submit" formmethod="get" formaction="EditTime"  value="@Resources.Resource.Edit" class="btn btn-primary c2" />
-                    <input type="submit" formaction="DeleteTime" value="Удалить время" class="btn btn-primary c2 " /><br /><br />
+                <form asp-controller="Time" method="post" style="border:0px;">
+                    <label class="whiteY" style="text-align:center;">@Resources.Resource.SavT</label><br />
+             @*    <select asp-for="Id" size="14" class="form-control card _c7" asp-items="ViewBag.TimeId"></select> *@
+                    <select asp-for=@Model.Time.Id size="10"  name="timeId" class="idTime form-control card _c7" asp-items="ViewBag.TimeId"></select>
+                    <br />
+                <div class="form-group">
+                        <input type="submit" formmethod="get" formaction="EditTime" value="@Resources.Resource.Edit" class="btn btn-primary  buttonsGY" /><br /><br />
+                        
                 </div>
+                
+                    <a id="btnDelTime" class="btn btn-primary   buttonsGY ">@Resources.Resource.Delete</a><br />
             </form>
         </div>
     </div>
+  </div>
+  <div >
+        <h4 class="whiteY" style="text-align:center;padding:5px;box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black;">
+            @Resources.Resource.MakeCh
+        </h4>
+
+    <div class="flex padd div_reg2">
+
+        <div class="padd">
+                <h4 class="whiteY">@Resources.Resource.ChDay</h4>
+                <form asp-controller="Time" method="post" style=" background-color: #2F4F4F;padding-top:30px;">
+              @*   @model SportClub.BLL.DTO.TimetableDTO *@
+
+                    <select asp-for=@Model.Timetable.Id size="10" class="form-control card _c7" asp-items="ViewBag.TimetableId" ></select>
+                <br />
+                <div class="form-group">
+                        <input type="submit" formaction="AddTimetable" value=@Resources.Resource.saveSc class="btn btn-primary buttonsGY" /><br /><br />
+                      </div>
+            </form>
+                <a asp-area="" asp-controller="Time" asp-action="Cancel" class="btn btn-primary buttonsGY">@Resources.Resource.Cancel</a>
+               @*  <a asp-area="" asp-controller="Time" asp-action="Exit" class="btn btn-primary buttonsGY">выход</a> *@
+
+        </div>
+        <div style="width:20px;"></div>
+        <div class="padd">
+           
+            <form asp-controller="Time" method="post">
+                    <label class="whiteY" style="text-align:center;">@Resources.Resource.chooseTime</label><br />
+                <br />
+               @*  <select asp-for="Id" size="10" class="form-control card _c7" asp-items="ViewBag.TimeId"></select> *@
+                     <select asp-for=@Model.Time.Id size="7" name="Tid" class="form-control card _c7" asp-items="ViewBag.TimeId"></select> 
+                <br /> <br />
+                <div class="form-group">
+                        <input type="submit" formmethod="get" formaction="AddTimesToTable" value=@Resources.Resource.addTtoCh class="btn btn-primary2  buttonsGY" />
+                   
+                </div>
+                    <br />
+                    <a href="#all" class="btn btn-primary buttonsGY">@Resources.Resource.AllSc</a>
+            </form>
+        </div>
+    </div>
+  </div>
+
 </div>
+<div style="padding-top:30px;">
+    <h4 id="all" class="whiteY" style="text-align:center;padding:5px;box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black;">@Resources.Resource.AllSc:</h4>
+</div>
+<div class="flex padd div_reg3" style="flex-wrap:wrap;background-color:  #2F4F4F;padding-top:0px;justify-content:space-around;">
+    @foreach (var item in Model.times)
+    {
+        <div class="padd" style="border-radius:3px;">
+            <form method="get">
+                <div class="padd" style="background-color:dimgray;border-radius:3px;">
+                    <input type="hidden"  name="id" value="@item.Id" />
+                    @foreach (var t in item.Times)
+                    {
+                        <input type="text" value="@t" style="width:150px;"> <br />
+                    }
+                    <br />
+                   @*  <input type="submit" style="margin-left:10px;" formaction="DeleteTimetable" readonly value="Удалить" class="btn btn-primary c2" /> *@
+                   @{var s="btn"+@item.Id;}
+                    <a id=@s class="btn btn-primary buttonsGY btnDelTimeTable">@Resources.Resource.Delete</a>
+                </div>
+            </form>
+        </div>
+    }
+    <form class="padd" >
+        <div class="padd" style="background-color:  dimgray;border-radius:3px;">
+            <input value=@Resources.Resource.oDay readonly style="width:120px;" />
+        </div>
+    </form>
+
+</div>
+
+@section Scripts {
+    @{
+        await Html.RenderPartialAsync("_ValidationScriptsPartial");
+    }
+}
+<script src="~/lib/jquery/dist/jquery.min.js"></script>
+<script>
+    $(document).ready(function () {
+
+        $("#start").on("change", function (e) {
+            let value = e.target.value;
+            let pattern="^(0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]$";
+            let r = true;
+            if (Regex.IsMatch(value, pattern)) {
+                        $("#st").css("display", "block");
+                        r = false;
+            }
+            if (r) {
+                $("#st").css("display", "none");
+            }
+        });
+        $("#end").on("change", function (e) {
+            let value = e.target.value;
+            let pattern = "^(0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]$";
+            let r = true;
+            if (Regex.IsMatch(value, pattern)) {
+                $("#en").css("display", "block");
+                r = false;
+            }
+            if (r) {
+                $("#en").css("display", "none");
+            }
+        });
+        var idTi;
+        var selectElement = document.querySelector('.idTime');    
+        selectElement.addEventListener('change', function () {
+            idTi = this.value; 
+        });
+
+       
+
+        $("#btnDelTime").click(function () { 
+            
+                $("<div id='dialogContent3'></div>")
+                    .addClass("dialog")
+                    .appendTo("body")
+                   .load("Admin/AddedPost")
+                    .dialog({
+                    title: "Удалить?",
+                        close: function () { $(this).remove(); },
+                        dialogClass: 'myTitleClass',
+                        modal: true,
+                        resizable: false,
+                        buttons: [{
+                            class: "input_subm3",//
+                            text: "Удалить",
+                            click: function () {
+                                $.ajax({
+                                    url: "@Url.Action("ConfirmDeleteTime", "Time")",
+                                    type: "POST",
+                                    data: { Id: idTi },
+                                    datatype: "json",
+                                    success: function (response) {
+
+                                        if (response) {
+                                            $("#dialogContent3").remove();
+                                            $("<div id='dialogContent4'></div>")
+                                                .addClass("dialog")
+                                                .appendTo("body")
+                                                .load("/Admin/AddedPost")
+                                                .dialog({
+                                                    title: "Удалено!",
+                                                    close: function () { window.location.href = "@Url.Action("AddTimeT", "Time")"; },
+                                                    dialogClass: 'myTitleClass',
+                                                    modal: true,
+                                                    resizable: false,
+                                                    buttons: [{
+                                                        class: "input_subm2",//
+                                                        text: "Закрыть",
+                                                        click: function () {
+                                                        window.location.href = "@Url.Action("AddTimeT", "Time")";
+                                                        }
+                                                    }],
+                                                    open: function (event, ui) {
+                                                        StyleDialog($(this).closest(".ui-dialog"));
+                                                    }
+                                                });
+                                        }
+                                        else {
+                                            $("#dialogContent3").remove();
+                                            $("<div id='dialogContent4'></div>")
+                                                .addClass("dialog")
+                                                .appendTo("body")
+                                                .load("/Admin/AddedPost")
+                                                .dialog({
+                                                    title: "Ошибка удаления!",
+                                                    close: function () { window.location.href = "@Url.Action("AddTimeT", "Time")"; },
+                                                    dialogClass: 'myTitleClass',
+                                                    modal: true,
+                                                    resizable: false,
+                                                    buttons: [{
+                                                        class: "input_subm3",//
+                                                        text: "Закрыть",
+                                                        click: function () {
+                                                        window.location.href = "@Url.Action("AddTimeT", "Time")";
+                                                        }
+                                                    }],
+                                                    open: function (event, ui) {
+                                                        StyleDialog($(this).closest(".ui-dialog"));
+                                                    }
+                                                });
+                                        }
+                                    }
+                                });
+                            }
+                        }],
+                        open: function (event, ui) {
+                        $(this).closest(".ui-dialog")
+                            .find(".ui-dialog-titlebar")
+                            .css({
+                                "background-color": "rgb(80, 80, 80)", // Фон заголовка
+                                "color": "#EEE8AA"
+                            });
+                        $(this).closest(".ui-dialog")
+                            .css({
+                                "box-shadow": "inset 0 - 3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black",
+                                "border-radius": "5px",
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+
+                            });
+                        $(this).closest(".ui-dialog")
+                            .find(".ui-dialog-content")
+                            .css({
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+                            });
+                        $(this).closest(".ui-widget-content")
+                            .css({
+                                "border": "none",
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+                            });
+                        $(this).closest(".ui-dialog")
+                            .find(".ui-dialog-buttonpane")
+                            .css({
+                                "border-width": "0px",
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+                            });
+                        }
+                    });         
+        });
+
+        $(".btnDelTimeTable").click(function () {
+            var idd = $(this).attr('id');
+            var result = idd.match(/\d+/);
+            $("<div id='dialogContent3'></div>")
+                .addClass("dialog")
+                .appendTo("body")
+                .load("Admin/AddedPost")
+                .dialog({
+                    title: "Удалить?",
+                    close: function () { $(this).remove(); },
+                    dialogClass: 'myTitleClass',
+                    modal: true,
+                    resizable: false,
+                    buttons: [{
+                        class: "input_subm3",//
+                        text: "Удалить",
+                        click: function () {
+                            $.ajax({
+                                url: "@Url.Action("ConfirmDeleteTimetable", "Time")",
+                                type: "POST",
+                                data: { Id: result },
+                                datatype: "json",
+                                success: function (response) {
+
+                                    if (response) {
+                                        $("#dialogContent3").remove();
+                                        $("<div id='dialogContent4'></div>")
+                                            .addClass("dialog")
+                                            .appendTo("body")
+                                            .load("/Admin/AddedPost")
+                                            .dialog({
+                                                title: "Удалено!",
+                                                close: function () { window.location.href = "@Url.Action("AddTimeT", "Time")"; },
+                                                dialogClass: 'myTitleClass',
+                                                modal: true,
+                                                resizable: false,
+                                                buttons: [{
+                                                    class: "input_subm2",//
+                                                    text: "Закрыть",
+                                                    click: function () {
+                                                        window.location.href = "@Url.Action("AddTimeT", "Time")";
+                                                    }
+                                                }],
+                                                open: function (event, ui) {
+                                                    StyleDialog($(this).closest(".ui-dialog"));
+                                                }
+                                            });
+                                    }
+                                    else {
+                                        $("#dialogContent3").remove();
+                                        $("<div id='dialogContent4'></div>")
+                                            .addClass("dialog")
+                                            .appendTo("body")
+                                            .load("/Admin/AddedPost")
+                                            .dialog({
+                                                title: "Ошибка удаления!",
+                                                close: function () { window.location.href = "@Url.Action("AddTimeT", "Time")"; },
+                                                dialogClass: 'myTitleClass',
+                                                modal: true,
+                                                resizable: false,
+                                                buttons: [{
+                                                    class: "input_subm3",//
+                                                    text: "Закрыть",
+                                                    click: function () {
+                                                        window.location.href = "@Url.Action("AddTimeT", "Time")";
+                                                    }
+                                                }],
+                                                open: function (event, ui) {
+                                                    StyleDialog($(this).closest(".ui-dialog"));
+                                                }
+                                            });
+                                    }
+                                }
+                            });
+                        }
+                    }],
+                    open: function (event, ui) {
+                        $(this).closest(".ui-dialog")
+                            .find(".ui-dialog-titlebar")
+                            .css({
+                                "background-color": "rgb(80, 80, 80)", // Фон заголовка
+                                "color": "#EEE8AA"
+                            });
+                        $(this).closest(".ui-dialog")
+                            .css({
+                                "box-shadow": "inset 0 - 3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black",
+                                "border-radius": "5px",
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+
+                            });
+                        $(this).closest(".ui-dialog")
+                            .find(".ui-dialog-content")
+                            .css({
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+                            });
+                        $(this).closest(".ui-widget-content")
+                            .css({
+                                "border": "none",
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+                            });
+                        $(this).closest(".ui-dialog")
+                            .find(".ui-dialog-buttonpane")
+                            .css({
+                                "border-width": "0px",
+                                "background-color": "rgb(80, 80, 80)",
+                                "background - image": "url('~/Coaches/фон5.jpg')"
+                            });
+                    }
+                });
+
+            
+
+        });
+
+    });
+</script>
