@@ -8,7 +8,7 @@
         <input type="hidden" name="time" value="@Model.Time" readonly /><br />
         <input type="hidden" name="roomId" value="@Model.RoomId" />
         <input type="hidden" name="roomName" value="@Model.RoomName" readonly /><br />
-        <input type="hidden" name="coachName" value="@Model.RoomName" readonly /><br />
+        <input type="hidden" name="coachName" value="@Model.CoachName" readonly /><br />
         <label>Выбрать клиента</label>
         <select asp-for="Id" name="userId" class="form-control card _c7" asp-items="ViewBag.UserId"></select>
 
