@@ -1,8 +1,8 @@
-﻿<div class="center pr">
+﻿@* <div class="div_reg padd">
 
 
-    <div class=" card c7">
-        <h4>@Resources.Resource.Addition</h4>
+    <div class=" card c7 padd">
+        <h4>@Resources.Resource.Delete</h4>
         <form asp-controller="Time" method="post">
             @model SportClub.BLL.DTO.TimeTDTO
             <input type="hidden" asp-for=Id />
@@ -16,10 +16,15 @@
                 <span>Время окончания</span><br />
                 <input asp-for="EndTime" name="End" class="form-control card _c7" readonly />
                 <span asp-validation-for="EndTime" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <input type="submit" formaction="ConfirmDeleteTime" value="Потвердить удаление" class="btn btn-primary c2" />
+            </div>  
+    <br />
+    <h5 style="color:forestgreen">Время будет удалено во всех графиках и расписаниях!</h5>
+
+    <div class="form-group">
+                <input type="submit" formaction="ConfirmDeleteTime" value="Потвердить удаление" class="btn btn-primary c2 buttons" />
+                <input type="submit" formaction="ExitTime" value="Отмена" class="btn btn-primary c2 buttons" />
+
             </div>
         </form>
     </div>
-</div>
+</div> *@
