@@ -1,26 +1,36 @@
-﻿<div class="center pr">
-   
+﻿<div class="div_reg padd">
 
-        <div class=" card c7">
-            <h4>@Resources.Resource.Addition</h4>
+   
+        <div class=" card c7 padd">
+        <h4>@Resources.Resource.Edit</h4>
             <form asp-controller="Time" method="post">
-                @model SportClub.BLL.DTO.TimeTDTO
-                <input type="hidden" asp-for=Id />
-                <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+            @model SportClub.BLL.DTO.TimeTDTO
+                        
+                <input type="hidden" asp-for=Id />              
+
+            @if (@ViewBag.ErrorMessage != null)
+            {
+                <div class="text-danger">@ViewBag.ErrorMessage</div>
+            }
+           
                 <div class="form-group">
                     <span>Время начала</span><br />
                     <input asp-for="StartTime" name="Start"  class="form-control card _c7" />
-                    <span asp-validation-for="StartTime" class="text-danger"></span>
+                  
                 </div>
                 <div class="form-group">
                     <span>Время окончания</span><br />
                     <input asp-for="EndTime"  name="End" class="form-control card _c7" />
-                    <span asp-validation-for="EndTime" class="text-danger"></span>
+                  
                 </div>
+                <br />
+            <h5 style="color:forestgreen">Время будет отредактировано во всех графиках и расписаниях!</h5>
                 <div class="form-group">
-                    <input type="submit" formaction="EditTime" value="Изменить" class="btn btn-primary c2" />
-                    <input type="submit" formaction="Exit" value="Отмена" class="btn btn-primary c2" />
+                    <input type="submit" formaction="EditTime" value="Изменить" class="btn btn-primary c2 buttons" />
+                    <input type="submit" formaction="ExitTime" value="Отмена" class="btn btn-primary c2 buttons" />
                 </div>
             </form>
         </div>
-    </div>
+
+
+</div>
