@@ -4,12 +4,12 @@
     ViewData["Title"] = "Timetables";
 }
 
-<h4>Графики:</h4>
+<h4>Добавьте один из графиков в расписаиние:</h4>
 <div class="flex">
     
     @foreach (var item in Model.times)
     {
-        <div>
+        <div class="padd1 marg">
           <form method="post">
                <input type="hidden" name="roomId" value="@Model.room.Id" />
               <input type="hidden" name="id" value="@item.Id" />
@@ -22,11 +22,13 @@
         </div>
     }
     <form method="post">
+        <div class="padd1 marg">
     <input value="ВЫХОДНОЙ" readonly /><br />
         <input type="hidden" name="roomId" value="@Model.room.Id" />
         <input type="hidden" name="id" value=0 />
     <input type="submit" formaction="AddTimetableToShedule" value="Добавить" class="btn btn-primary c2" />
-    </form>
+   </div>
+</form>
 </div>
 <div>
     <h1>@Model.room.Name</h1>
@@ -45,18 +47,22 @@
                 @foreach (var item in Model.timesAdded)
                 {
                     <td>
+                        <div class="padd1 marg">
                         <input type="hidden" name="tId" value="@item.Id" />
                                                                           
                             @foreach (var t in item.Times)
                             {
                                 <input type="text" value="@t" style="width:150px;"> <br />
                             }                         
-                       
+                       </div>
                     </td>
                 }
 </tr>
 </tbody>
     </table>
+    <a asp-area="" asp-route-roId="@Model.room.Id" asp-controller="Time" asp-action="CancelAddTimetable" class="btn btn-primary padd1 buttons1">Отмена</a>
+</div>
+
     <form method="post">
         <input type="hidden" name="rId" value="@Model.room.Id" />
         <input type="submit" asp-controller="Time" asp-action="SaveShedule" value="Сохранить график работы" class="btn btn-primary c2" />
