@@ -2,44 +2,234 @@
 @{
     ViewData["Title"] = "Timetables";
 }
-<form method="post">
-   
-        <h1>@Model.room.Name</h1>
-        <table style="border-collapse: collapse; border: 1px solid black;">
-            <thead>
-            <th style=" padding: 8px; border: 1px solid black;">Понедельник</th>
-            <th style=" padding: 8px; border: 1px solid black;">Вторник</th>
-            <th style=" padding: 8px; border: 1px solid black;">Среда</th>
-            <th style=" padding: 8px; border: 1px solid black;">Четверг</th>
-            <th style=" padding: 8px; border: 1px solid black;">Пятница</th>
-            <th style=" padding: 8px; border: 1px solid black;">Суббота</th>
-            <th style=" padding: 8px; border: 1px solid black;">Воскресенье</th>
-            </thead>
-            <tbody>              
-                <tr>
-                    @foreach (var item in Model.times)
-                    {
-                        <td>
 
-                            @for (var t=0;t<item.Times.Count;t++)
-                            {
-                            <input type="hidden" name="day" value="@t" />
-                            <input type="hidden" name="roomId" value="@Model.room.Id" />
-                            <input type="hidden" name="roomName" value="@Model.room.Name" />
-                           
-                            <input type="text" name="time" value="@item.Times[t]" style="width:100px;" readonly> <br />
-                            @if (@item.Times[t] != "Выходной")
-                            {
-                                <input type="submit" formaction="AddIndTraining" value="Добавить индивидуальную тренировку" /> <br />
-                                <input type="submit" formaction="AddGroupTraining" value="Добавить групповую тренировку" />  <br />
-                            }
+<div class=" center1 padd">
+    <div style="padding-right:80px">
+        <a asp-area="" asp-controller="Time" asp-action="BackToRooms" class="btn btn-primary padd1 buttons1">@Resources.Resource.Back</a>
+    </div>
+    <input type="hidden" id="idshed" value="@Model.room.sheduleId" />
+    <input type="hidden" id="idroom" value="@Model.room.Id" />
+    <h1>@Model.room.Name &nbsp;&nbsp;&nbsp;</h1>
+    <img src="@Model.room.Photo" style="width:70px;height:70px" alt="photo" />
+    @if (Model.message == null)
+    {
+        <a href="javascript:void(0)" id="btnDelShed" class="btn btn-primary padd1 buttons1" style="margin-left:80px;">
+            Удалить расписание зала
+        </a>
+    }
+</div>
+@if (Model.message == null)
+{
+    <table style=" border: 1px solid black; margin-left:-100px;">
+        <thead>
+            <tr>
+                <td><h6 class="padd1 white marg ">Понедельник</h6></td>
+                <td><h6 class="padd1 white marg ">Вторник </h6></td>
+                <td><h6 class="padd1 white marg ">Среда</h6></td>
+                <td><h6 class="padd1 white marg ">Четверг</h6></td>
+                <td><h6 class="padd1 white marg ">Пятница</h6></td>
+                <td><h6 class="padd1 white marg ">Суббота</h6></td>
+                <td><h6 class="padd1 white marg ">Воскресенье</h6></td>
+            </tr>
+        </thead>
+        <tbody>
+            <tr>
+                @for (var t1 = 0; t1 < Model.times.Count; t1++)
+                {
+
+                    <td>
+
+                        @for (var t = 0; t < Model.times[t1].Times.Count; t++)
+                        {
+
+                            <form method="post" class="padd1 marg">
+
+                                <input type="hidden" name="day" value="@t1" />
+                                <input type="hidden" name="roomId" value="@Model.room.Id" />
+                                <input type="hidden" name="roomName" value="@Model.room.Name" />
+
+                                <input type="text" name="time" value="@Model.times[t1].Times[t]" style="width:120px; background-color:yellow;font-size:20px;" readonly> <br />
+
+
+                                @foreach (var it in Model.trainingInd)
+                                {
+                                    @if (@Model.times[t1].Times[t] != "Выходной" && @it.Time == @Model.times[t1].Times[t] && @it.Day == @t1 && it.RoomId == @Model.room.Id)
+                                    {
+                                        Model.flag = true;
+                                        <div id=@t1+@Model.times[t1].Times[t]+@t1 class="flex">
+                                            <div>
+                                                <span class="white">Инд.тренировка</span><br />
+                                                <span class="white">Тренер : </span><br />
+                                                <span class="white">@it.CoachName</span><br />
+                                                @if(it.UserId == null)
+                                                {
+                                                    <a class="btn btn-outline-dark" style="background-color:lightgreen" asp-area="" asp-route-roomId="@it.RoomId" asp-route-groupId="@it.Id" asp-controller="TrainingInd" asp-action="Hello">Записаться!!!</a>
+                                                }
+                                                else
+                                                {
+                                                    <span class="white">@it.UserName</span><br />
+                                                    <a class="btn btn-outline-dark" style="background-color:coral" asp-area="" asp-route-roomId="@it.RoomId" asp-route-groupId="@it.Id" asp-controller="" asp-action="">Просмотр</a>
+                                                }
+                                            </div>
+                                            <img src="@it.CoachPhoto" style="width:50px;height:50px" alt="photo" class="marg" />
+                                        </div>
+                                    }
+                                }
+                                @foreach (var it in Model.traininggroup)
+                                {
+                                    @if (@Model.times[t1].Times[t] != "Выходной" && @it.Time == @Model.times[t1].Times[t] && @it.Day == @t1 && it.Room.Id == @Model.room.Id)
+                                    {
+                                        Model.flag = true;
+                                        <div id=@t1+@Model.times[t1].Times[t]+@t1 class="flex">
+                                            <div style="min-width:190px">
+                                                @{
+                                                    var num = it/* .Group */.Number - it.Users.Count;
+                                                }
+                                                <span class="white">Места: @it.Number</span><br />
+                                                @* <span class="white">@it.Group.Name ( @it.Group.Number )</span><br /> *@
+                                                <span class="white">Тренер : </span><br />
+                                                <span class="white">@it.Coach.Name</span><br />
+                                                @if (it.Number > it.Users.Count)@* (it.Group.Number>it.Group.UsersId.Count) *@
+                                                {<span class="white">Доступно : @num </span><br />
+                                                    <div style="margin: 0px auto 0px auto;width: max-content;">
+                                                        <a class="btn btn-outline-dark" style="background-color:lightgreen" asp-area="" asp-route-roomId="@it.Room.Id" asp-route-trgroupId="@it.Id" asp-controller="Users" asp-action="AddUsersToTrainingGroup">Записать</a>
+                                                        <a class="btn btn-outline-dark" style="background-color:coral; " asp-area="" asp-route-Id="@it.Id" asp-controller="Time" asp-action="DetailsTrainingGroup">Просмотр</a>
+                                                    </div>
+                                                    
+                                                    
+                                                }
+                                                else
+                                                {
+                                                    <span class=" white"> заполнена</span>
+
+                                                    <br /> <a class="btn btn-outline-dark" style="background-color:coral" asp-area="" asp-route-Id="@it.Id" asp-controller="Time" asp-action="DetailsTrainingGroup">Просмотр</a>
+                                                }@* asp-route-roomId="@it.Room.Id"*@
+                                            </div>
+                                            <img src="@it.Coach.Photo" style="width:50px;height:50px;margin-left:-50px; margin-top:-35px" alt="photo" class="marg" />
+                                        </div>
+                                    }
+                                }
+
+                                @if (@Model.times[t1].Times[t] != "Выходной" && !@Model.flag)
+                                {
+                                    <div id=@t1+@Model.times[t1].Times[t]>
+                                        <input type="submit" formaction="AddIndTraining" value="Добавить и.тренировку" /> <br />
+                                        <input type="submit" formaction="AddTrainingGroup" value="Добавить г.тренировку" />  <br />
+                                    </div>
+
+                                }
+
+
+                            </form>
+
+                            Model.flag = false;
                         }
+                        @*    <a asp-route-SheduleId="@Model.room.sheduleId" asp-route-day="@t1" asp-route-RoomId="@Model.room.Id" asp-controller="Time" asp-action="GetTimetableToEdit" class="btn btn-primary padd1 buttons1">Редактировать график</a> *@
 
-                        </td>
+                    </td>
+
+                }
+
+            </tr>
+
+        </tbody>
+    </table>
+}
+@if (Model.message != null)
+{
+    <div class="center">
+        <div style="display:flex;flex-direction:column;">
+            <div>
+                <h1>@Model.message</h1><div style="display:flex;flex-direction:column;">
+                </div><br /><br />
+                <div>
+                    <form method="post">
+
+                        <input type="hidden" name="Id" value="@Model.room.Id" />
+                        <input type="submit" asp-controller="Time" asp-action="GetAllTimetable" class="padd1 white" value="Составить расписание зала" />
+
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+}
+
+<script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
+<script>
+    $(document).ready(function () {
+
+        $("#btnDelShed").click(function () {
+            // var idS = $("#idshed").val();
+            var idR = $("#idroom").val();
+            var puth = "/Time/DeleteShedule/?RomId=" + idR;
+            $("<div id='dialogContent3'></div>")
+                .addClass("dialog")
+                .appendTo("body")
+                .load(puth)
+                .dialog({
+                    title: "Подтвердите!",
+                    close: function () { $(this).remove(); },
+                    dialogClass: 'myTitleClass',
+                    modal: true,
+                    resizable: false,
+                    buttons: [{
+                        class: "input_subm2",
+                        text: "Удалить",
+                        click: function () {
+                            $.ajax({
+                                url: "@Url.Action("DeleteSheduleInRoom", "Time")",
+                                type: "POST",
+                                data: { RomId: idR },
+                                datatype: "json",
+                                success: function (response) {
+
+                                    if (response) {
+
+                                        window.location.href = "@Url.Action("Room_Shedule", "Time")";
+                                    }
+                                }
+                            });
+                        }
+                    }],
+                    open: function (event, ui) {
+                        StyleDialog($(this).closest(".ui-dialog"));
                     }
-                </tr>
+                });
 
-            </tbody>
-        </table>
-       <h1>@Model.message</h1>
-</form>
+        });
+
+        function StyleDialog(dialog) {
+            dialog
+                .find(".ui-dialog-titlebar")
+                .css({
+                    "background-color": "#374566", // Фон заголовка
+                    "color": "white" // Цвет текста заголовка black
+                });
+            dialog
+                .css({
+                    "box-shadow": "inset 0 - 3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black",
+                    "border-radius": "10px",
+                    "border": "solid 1px #374566",
+                    "background-color": "#374566"
+                });
+            dialog
+                .find(".ui-dialog-content")
+                .css({
+                    "background-color": "#374566"
+                });
+            dialog
+                .find(".ui-widget-content")
+                .css({
+                    "border": "none",
+                    "background-color": "#374566"
+                });
+            dialog
+                .find(".ui-dialog-buttonpane")
+                .css({
+                    "border-width": "0px",
+                    "background-color": "#374566"
+                });
+        }
+    });
+</script>
