@@ -1,9 +1,27 @@
-﻿@model SportClub.BLL.DTO.RoomDTO
+﻿@* @model SportClub.BLL.DTO.RoomDTO
 <form method="post">
     <div>
         <h3>Посмотреть расписание зала</h3>
         <h4>Выберите спортзал</h4><br />
-        <select asp-for="Id" style="width:200px;" class="form-control card _c7" asp-items="ViewBag.RoomsId"></select>
+        <select asp-for="Id" name="RoomId" style="width:200px;" class="form-control card _c7" asp-items="ViewBag.RoomsId"></select>
     </div><br />
     <input type="submit" asp-controller="Time" asp-action="RoomWithShedule" value="Выбрать" />
-</form>
+</form> *@
+
+@model IEnumerable<SportClub.BLL.DTO.RoomDTO>
+<div class="center"><h2>Расписание</h2></div>
+<div class="flex center" style="flex-wrap:wrap;">
+@foreach (var item in Model)
+{
+    <div class="padd1" style="margin:10px;">
+      <form method="post">
+        <img src=@item.Photo width="200" height="200" alt="Photo" style="background-color:wheat" />
+        <div>
+            <h4 class="white"> @item.Name</h4>
+        </div>
+                <input type="hidden" value="@item.Id" name="RoomId" />
+    <input type="submit" asp-controller="Time" asp-action="RoomWithShedule" value="Выбрать" />
+      </form>
+    </div>
+}
+</div>
