@@ -1,10 +1,11 @@
 ﻿@model SportClub.Models.CangePasswordModel
 @using Microsoft.AspNetCore.Http
+  
+   @*      <form class="card c22 login">
+    <div style="background-color: #374566; box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black; border-radius: 15px; width: 300px; margin-left: 400px; padding:20px">
 
-    <div class="center ">
-        <div>
-        <form class="card c22 login">
-            @Html.AntiForgeryToken()
+
+        @Html.AntiForgeryToken()
             <input type="hidden" asp-for="Id"  />
                 <span>Введите новый пароль</span><br />
             <input type="password" asp-for="Password" /><br />
@@ -15,10 +16,30 @@
                     <div>
                         <input type="submit" class="y" formaction="SaveNewPassword" value="Изменить пароль" />
                     </div>
-                          
-        </form>
-        </div>
     </div>
+</form>
+        *@
+<div style="background-color: #374566; box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0.3em 0.3em 1em black; border-radius: 15px; width: 300px; margin-left: 400px; padding:20px">
+    <form>
+        @Html.AntiForgeryToken()
+        <input type="hidden" asp-for="Id" />
+        <div class="form-group" style="margin-left: 20px">
+            <span style="color: white" class="font form-check-label">Введите новый пароль</span><br />
+            <input type="password" asp-for="Password" class="form-control" /><br />
+            <span asp-validation-for="Password" class="text-danger"></span>
+        </div>
+        <div class="form-group" style="margin-left: 20px">
+            <span style="color: white" class="font form-check-label">Подтвердите пароль</span><br />
+            <input type="password" asp-for="PasswordConfirm" class="form-control" /><br />
+            <span asp-validation-for="PasswordConfirm" class="text-danger"></span>
+        </div>
+        <div>
+            <input type="submit" class="btn btn-outline-light" formaction="SaveNewPassword" value="Изменить пароль" />
+        </div>
+    </form>
+</div>
+
+
 @section Scripts {
     @{
         await Html.RenderPartialAsync("_ValidationScriptsPartial");
