@@ -1,77 +1,57 @@
-﻿@model SportClub.BLL.DTO.CoachDTO
-
-
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
 @{
-    ViewData["Title"] = "Delete";
 }
+@model SportClub.BLL.DTO.UserDTO
+<div class="div_reg">
+    <p class="font" style="font-size: 32px; color: white; padding-top: 30px; text-align: center">
+        Удаление пользователя </p>
+    <p class="font" style="font-size: 22px; color: white; text-align: center"> Вы действительно желаете удалить<br /> пользователя? </p>
 
-<h1>Delete</h1>
+    <div class="row">
+        <div>
+            
+            <div class="row">
+                <div style="width: 500px;  justify-content: center; ">
+                    <div class="form-group" style="margin-left: 40px">
+                        <label asp-for="Name" class="control-label font  white-color_size20" style="color: white"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Name)</h1>
+                    </div>
+                    <div class="form-group" style="margin-left: 40px">
 
-<h3>Are you sure you want to delete this?</h3>
-<div>
-    <h4>User</h4>
-    <hr />
-    <dl class="row">
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Name)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Name)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.DateOfBirth)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.DateOfBirth)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Phone)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Phone)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Email)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Email)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Gender)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Gender)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Description)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Description)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Photo)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Photo)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Login)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Login)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Password)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Password)
-        </dd>
-       
-    </dl>
-    
-    <form asp-action="Delete">
-        <input type="hidden" asp-for="Id" />
-        <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-action="Index">Back to List</a>
-    </form>
+                        <label asp-for="Login" class="control-label font  white-color_size20" style="color: white"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Login)</h1>
+                    </div>
+
+                    <div class="form-group" style="margin-left: 40px">
+                        <label asp-for="Email" class="control-label font  white-color_size20" style="color: white"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Email)</h1>
+                    </div>
+
+                    <div class="form-group" style="margin-left: 40px">
+                        <label asp-for="Phone" style="color: white" class="control-label font  white-color_size20"></label>
+                        <h1 class="form-control p1" style="width: 420px">@Html.DisplayFor(model => model.Phone)</h1>
+                    </div>
+                </div>
+
+                <div class="form-group" style="margin-left: 40px">
+                    <label asp-for="DateOfBirth" class="control-label font  white-color_size20" style="color: white"></label>
+                    <h1 class="form-control p1" style="width: 420px;height: 33px">@Html.DisplayFor(model => model.DateOfBirth)</h1>
+                </div> <br />
+ 
+
+                <div style="margin: 10px auto 10px auto;width: max-content;">
+                    <form method="post" asp-action="Delete" asp-controller="Users"
+                          style="width: max-content; float: left;height: 38px;">
+                        <input type="hidden" asp-for="Id" name="id" />
+                        <input class="input_subm_del font y" style="height: 38px;" type="submit" value="УДАЛИТЬ" />
+                    </form>
+                    <a asp-action="GetClients" asp-controller="Users" style="margin-left: 10px;" class="btn btn-outline-light">Back to List</a>
+                </div><br />
+            </div>
+
+        </div>
+    </div>
 </div>
+ 
